Some refactoring

+++ src/engine/core/Input.cpp
@@ -186,8 +186,8 @@ bool Input::remove_input_event_handler(InputEventHandlerId id) {
         return false;
     }
 
-    int index = it->second;
-    int last_index = m_handlers.size() - 1;
+    size_t index = it->second;
+    size_t last_index = m_handlers.size() - 1;
 
     if (index != last_index) {
         m_handlers[index] = std::move(m_handlers[last_index]); // move last into hole
